pyproject.toml
[project]
name = "customer-service-ticket-triage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "kagglehub[polars-datasets]>=0.3.13",
    "outlines>=1.2.3",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "transformers>=4.55.4",
]
